Refactor the insights-service code to have a separate file for endpoint functions in addition to removing the logs endpoint and creating separate endpoints for getting logs based on pod name, namespace name, workload name and control plane component
